+++ changelog.md
@@ -16,6 +16,10 @@
 - scorelist filter args are rounded
 - scorelist filter args accept !<x> (filter by values that dont equal that number)
 
+### Added
+
+- support for mod settings (eg. difficulty adjust)
+
 ## [4.0.5] - 2024-12-18
 
 [commit](https://github.com/sbrstrkkdwmdr/sbrbot/commit/5c36d8cd93b03d25a0b2a537506ede93f8e43057)</br>